+++ app/Console/Kernel.php
@@ -13,8 +13,7 @@ class Kernel extends ConsoleKernel
 
     protected function schedule(Schedule $schedule)
     {
-        // $schedule->command('inspire')
-        //     ->hourly();
+        $schedule->command('telescope:prune')->daily();
     }
 
     protected function commands()