+++ app/Providers/TelescopeServiceProvider.php
@@ -14,14 +14,14 @@ class TelescopeServiceProvider extends TelescopeApplicationServiceProvider
         Telescope::night();
 
         Telescope::filter(function (IncomingEntry $entry) {
-            if ($this->app->environment('local')) {
+            if ($this->app->isLocal()) {
                 return true;
             }
 
             return $entry->isReportableException() ||
-                   $entry->isFailedJob() ||
-                   $entry->isScheduledTask() ||
-                   $entry->hasMonitoredTag();
+                $entry->isFailedJob() ||
+                $entry->isScheduledTask() ||
+                $entry->hasMonitoredTag();
         });
     }