+++ resources/js/components/enso/bulma/DateIntervalFilter.vue
@@ -1,26 +1,30 @@
 <template>
-    <div class="date-interval-filter">
-        <div class="has-text-centered">
-            <strong>{{ title }}</strong>
+    <div class="date-interval-filter is-paddingless">
+        <div v-if="!compact"
+            class="header has-background-light has-text-centered">
+            <strong>{{ i18n(title) }}</strong>
         </div>
-        <div class="columns is-mobile">
-            <div class="column">
-                <datepicker v-model="interval.min"
-                    :format="format"
-                    :is-warning="equal"
-                    :is-danger="inversed"
-                    :locale="locale"
-                    :placeholder="minLabel"
-                    @input="update"/>
-            </div>
-            <div class="column">
-                <datepicker v-model="interval.max"
-                    :format="format"
-                    :is-danger="inversed"
-                    :is-warning="equal"
-                    :locale="locale"
-                    :placeholder="maxLabel"
-                    @input="update"/>
+        <div v-tooltip="compact ? i18n(title) : null"
+            :class="['dates-wrapper', {'has-background-light': compact}]">
+            <div class="columns is-mobile is-variable is-1 is-centered">
+                <div class="column">
+                    <datepicker v-model="interval.min"
+                        :format="format"
+                        :is-warning="equal"
+                        :is-danger="inversed"
+                        :locale="locale"
+                        :placeholder="i18n(minLabel)"
+                        @input="update"/>
+                </div>
+                <div class="column">
+                    <datepicker v-model="interval.max"
+                        :format="format"
+                        :is-danger="inversed"
+                        :is-warning="equal"
+                        :locale="locale"
+                        :placeholder="i18n(maxLabel)"
+                        @input="update"/>
+                </div>
             </div>
         </div>
     </div>
@@ -29,17 +33,33 @@
 <script>
 
 import { compareAsc, parse } from 'date-fns';
+import { VTooltip } from 'v-tooltip';
 import Datepicker from '../vueforms/Datepicker.vue';
 
 export default {
     name: 'DateIntervalFilter',
 
+    directives: { tooltip: VTooltip },
+
     components: { Datepicker },
 
     props: {
+        compact: {
+            type: Boolean,
+            default: false,
+        },
         format: {
             type: String,
             default: 'd-m-Y',
+        },
+        i18n: {
+            type: Function,
+            default(key) {
+                return this.$options.methods
+                    && Object.keys(this.$options.methods).includes('__')
+                    ? this.__(key)
+                    : key;
+            },
         },
         locale: {
             type: String,
@@ -105,10 +125,19 @@ export default {
 
 </script>
 
-<style scoped>
+<style lang="scss" scoped>
 
     .date-interval-filter {
-        padding: 0.5rem;
+        .header {
+            border-top-left-radius: inherit;
+            border-top-right-radius: inherit;
+            padding-top: 0.5em;
+        }
+
+        .dates-wrapper {
+            border-radius: inherit;
+            padding: 0.25em;
+        }
     }
 
 </style>